Move `ast.ts` to `language` folder
