+++ packages/langium-cli/src/generator/module-generator.ts
@@ -22,7 +22,8 @@ export function generateModule(grammars: Grammar[], config: LangiumConfig, gramm
     ).appendIf(!!config.langiumInternal,
         expandToNode`
 
-            import type { LanguageMetaData } from '../language-meta-data${config.importExtension}';
+            import type { LanguageMetaData } from '../../language/language-meta-data${config.importExtension}';
+            import { ${config.projectName}AstReflection } from '../../language/generated/ast${config.importExtension}';
             import type { Module } from '../../dependency-injection${config.importExtension}';
             import type { LangiumGeneratedServices, LangiumGeneratedSharedServices, LangiumSharedServices, LangiumServices } from '../../services${config.importExtension}';
         `.appendTemplateIf(hasIParserConfigImport)`
@@ -32,9 +33,9 @@ export function generateModule(grammars: Grammar[], config: LangiumConfig, gramm
     ).appendTemplateIf(!config.langiumInternal)`
 
         import type { LangiumGeneratedServices, LangiumGeneratedSharedServices, LangiumSharedServices, LangiumServices, LanguageMetaData, Module${hasIParserConfigImport ? ', IParserConfig' : ''} } from 'langium';
+        import { ${config.projectName}AstReflection } from './ast${config.importExtension}';
     `.appendTemplate`
 
-        import { ${config.projectName}AstReflection } from './ast${config.importExtension}';
         import { ${joinToNode(grammarsWithName, grammar => grammar.name + 'Grammar', { separator: ', '}) } } from './grammar${config.importExtension}';
         ${joinToNode(
             grammarsWithName,
