+++ packages/langium/src/grammar/langium-grammar-module.ts
@@ -9,7 +9,7 @@ import type { Module } from '../dependency-injection.js';
 import type { LangiumServices, LangiumSharedServices, PartialLangiumServices, PartialLangiumSharedServices } from '../services.js';
 import { LangiumGrammarTypeHierarchyProvider } from './lsp/grammar-type-hierarchy.js';
 import type { LangiumGrammarDocument } from './workspace/documents.js';
-import type { Grammar } from './generated/ast.js';
+import type { Grammar } from '../language/generated/ast.js';
 import { createDefaultModule, createDefaultSharedModule } from '../default-module.js';
 import { inject } from '../dependency-injection.js';
 import { LangiumGrammarGeneratedModule, LangiumGrammarGeneratedSharedModule } from './generated/module.js';
