+++ packages/langium/src/grammar/lsp/grammar-semantic-tokens.ts
@@ -8,7 +8,7 @@ import type { AstNode } from '../../syntax-tree.js';
 import type { SemanticTokenAcceptor } from '../../lsp/semantic-token-provider.js';
 import { SemanticTokenTypes } from 'vscode-languageserver';
 import { AbstractSemanticTokenProvider } from '../../lsp/semantic-token-provider.js';
-import { isAction, isAssignment, isParameter, isParameterReference, isReturnType, isRuleCall, isSimpleType, isTypeAttribute } from '../generated/ast.js';
+import { isAction, isAssignment, isParameter, isParameterReference, isReturnType, isRuleCall, isSimpleType, isTypeAttribute } from '../../language/generated/ast.js';
 
 export class LangiumGrammarSemanticTokenProvider extends AbstractSemanticTokenProvider {
