+++ packages/langium/src/grammar/type-system/types-util.ts
@@ -7,11 +7,11 @@
 import type { References } from '../../references/references.js';
 import type { AstNodeLocator } from '../../workspace/ast-node-locator.js';
 import type { LangiumDocuments } from '../../workspace/documents.js';
-import type { Interface, Type, AbstractType, TypeDefinition } from '../generated/ast.js';
+import type { Interface, Type, AbstractType, TypeDefinition } from '../../language/generated/ast.js';
 import type { PlainInterface, PlainProperty } from './type-collector/plain-types.js';
 import type { AstTypes, InterfaceType, PropertyType, TypeOption } from './type-collector/types.js';
 import { MultiMap } from '../../utils/collections.js';
-import { isInterface, isType, isUnionType, isSimpleType } from '../generated/ast.js';
+import { isInterface, isType, isUnionType, isSimpleType } from '../../language/generated/ast.js';
 import { isArrayType, isPrimitiveType, isPropertyUnion, isReferenceType, isValueType } from './type-collector/types.js';
 
 /**
