+++ packages/langium/test/tsconfig.export-main.json
@@ -5,7 +5,8 @@
       "noEmit": true
     },
     "exclude": [
-        // "../src/generate/**/*", // cs: to be enabled once #1171 is merged
+        "../src/generate/**/*",
+        "../src/grammar/**/*",
         "../src/test/**/*"
     ]
 }
